+++ apps/storybook2/.storybook/StoryContainer.tsx
@@ -5,14 +5,15 @@ import { Story } from '@storybook/react';
 import merge from 'lodash/merge';
 import { useDarkMode } from 'storybook-dark-mode';
 import { sanitizeProps, StoryBuilderConfig } from '@cbhq/cds-common/internal/utils/storyBuilder';
-import { gutter } from '@cbhq/cds-common/tokens/sizing';
-import { Button } from '@cbhq/cds-web2/buttons/Button';
+// import { gutter } from '@cbhq/cds-common/tokens/sizing';
 
 import { PortalProvider } from '@cbhq/cds-web2/overlays/PortalProvider';
 import { ThemeProvider } from '@cbhq/cds-web2/providers/ThemeProvider';
 import { autoTheme } from '@cbhq/cds-web2/themes/auto';
 import { darkTheme } from '@cbhq/cds-web2/themes/dark';
 import { lightTheme } from '@cbhq/cds-web2/themes/light';
+// TODO migrate Group
+// import { Group } from '@cbhq/cds-web2/layout/Group';
 
 const themeMap = {
   auto: autoTheme,